requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0

# Downloaders
yt-dlp>=2023.7.6
requests>=2.31.0

# Audio processing
pydub>=0.25.1
fal-client>=0.3.0

# Google Vertex AI
google-cloud-aiplatform>=1.28.0
google-genai>=0.3.0

# Text processing and NLP
nltk>=3.8.1
langchain>=0.1.0

# Logging and utilities
colorlog>=6.7.0
tqdm>=4.66.0